+++ tests/e2e/websocket/common.hpp
@@ -14,6 +14,7 @@
 #include <cstdint>
 #include <iostream>
 #include <memory>
+#include <optional>
 #include <string>
 #include <string_view>
 #include <thread>
@@ -47,6 +48,8 @@ struct Credentials {
   std::string_view passsword;
 };
 
+inline void Fail(beast::error_code ec, char const *what) { std::cerr << what << ": " << ec.message() << "\n"; }
+
 std::string GetAuthenticationJSON(const Credentials &creds) {
   nlohmann::json json_creds;
   json_creds["username"] = creds.username;
@@ -54,7 +57,138 @@ std::string GetAuthenticationJSON(const Credentials &creds) {
   return json_creds.dump();
 }
 
-inline void Fail(beast::error_code ec, char const *what) { std::cerr << what << ": " << ec.message() << "\n"; }
+template <bool ssl = false>
+class Session : public std::enable_shared_from_this<Session<ssl>> {
+  using std::enable_shared_from_this<Session<ssl>>::shared_from_this;
+
+ public:
+  explicit Session(net::io_context &ioc, ssl::context &ctx, std::vector<std::string> &expected_messages) requires(ssl)
+      : resolver_(net::make_strand(ioc)), ws_(net::make_strand(ioc), ctx), received_messages_{expected_messages} {}
+
+  explicit Session(net::io_context &ioc, std::vector<std::string> &expected_messages) requires(!ssl)
+      : resolver_(net::make_strand(ioc)), ws_(net::make_strand(ioc)), received_messages_{expected_messages} {}
+
+  explicit Session(net::io_context &ioc, ssl::context &ctx, std::vector<std::string> &expected_messages,
+                   Credentials creds) requires(ssl)
+      : resolver_(net::make_strand(ioc)),
+        ws_(net::make_strand(ioc), ctx),
+        received_messages_{expected_messages},
+        creds_{creds} {}
+
+  explicit Session(net::io_context &ioc, std::vector<std::string> &expected_messages, Credentials creds) requires(!ssl)
+      : resolver_(net::make_strand(ioc)),
+        ws_(net::make_strand(ioc)),
+        received_messages_{expected_messages},
+        creds_{creds} {}
+
+  void Run(std::string host, std::string port) {
+    host_ = host;
+    resolver_.async_resolve(host, port, beast::bind_front_handler(&Session::OnResolve, shared_from_this()));
+  }
+
+  void OnResolve(beast::error_code ec, tcp::resolver::results_type results) {
+    if (ec) {
+      return Fail(ec, "resolve");
+    }
+
+    beast::get_lowest_layer(ws_).expires_after(std::chrono::seconds(30));
+    beast::get_lowest_layer(ws_).async_connect(results,
+                                               beast::bind_front_handler(&Session::OnConnect, shared_from_this()));
+  }
+
+  void OnConnect(beast::error_code ec, tcp::resolver::results_type::endpoint_type ep) {
+    if (ec) {
+      return Fail(ec, "connect");
+    }
+
+    host_ = fmt::format("{}:{}", host_, ep.port());
+
+    beast::get_lowest_layer(ws_).expires_after(std::chrono::seconds(30));
+
+    if constexpr (ssl) {
+      if (!SSL_set_tlsext_host_name(ws_.next_layer().native_handle(), host_.c_str())) {
+        ec = beast::error_code(static_cast<int>(::ERR_get_error()), net::error::get_ssl_category());
+        return Fail(ec, "connect");
+      }
+      ws_.next_layer().async_handshake(ssl::stream_base::client,
+                                       beast::bind_front_handler(&Session::OnSSLHandshake, shared_from_this()));
+    } else {
+      ws_.set_option(websocket::stream_base::timeout::suggested(beast::role_type::client));
+
+      ws_.set_option(websocket::stream_base::decorator([](websocket::request_type &req) {
+        req.set(http::field::user_agent, std::string(BOOST_BEAST_VERSION_STRING) + " websocket-client-async");
+      }));
+
+      host_ = fmt::format("{}:{}", host_, ep.port());
+
+      ws_.async_handshake(host_, "/", beast::bind_front_handler(&Session::OnHandshake, shared_from_this()));
+    }
+  }
+
+  void OnHandshake(beast::error_code ec) {
+    if (ec) {
+      return Fail(ec, "handshake");
+    }
+    if (creds_) {
+      ws_.async_write(net::buffer(GetAuthenticationJSON(*creds_)),
+                      beast::bind_front_handler(&Session::OnWrite, shared_from_this()));
+    } else {
+      ws_.async_read(buffer_, beast::bind_front_handler(&Session::OnRead, shared_from_this()));
+    }
+  }
+
+  void OnSSLHandshake(beast::error_code ec) {
+    if (ec) {
+      return Fail(ec, "ssl_handshake");
+    }
+
+    beast::get_lowest_layer(ws_).expires_never();
+    ws_.set_option(websocket::stream_base::timeout::suggested(beast::role_type::client));
+    ws_.set_option(websocket::stream_base::decorator([](websocket::request_type &req) {
+      req.set(http::field::user_agent, std::string(BOOST_BEAST_VERSION_STRING) + " websocket-client-async-ssl");
+    }));
+
+    ws_.async_handshake(host_, "/", beast::bind_front_handler(&Session::OnHandshake, shared_from_this()));
+  }
+
+  void OnWrite(beast::error_code ec, std::size_t bytes_transferred) {
+    boost::ignore_unused(bytes_transferred);
+
+    if (ec) {
+      return Fail(ec, "write");
+    }
+
+    ws_.async_read(buffer_, beast::bind_front_handler(&Session::OnRead, shared_from_this()));
+  }
+
+  void OnRead(beast::error_code ec, std::size_t bytes_transferred) {
+    boost::ignore_unused(bytes_transferred);
+
+    if (ec) {
+      return Fail(ec, "read");
+    }
+
+    received_messages_.push_back(boost::beast::buffers_to_string(buffer_.data()));
+    buffer_.clear();
+
+    ws_.async_read(buffer_, beast::bind_front_handler(&Session::OnRead, shared_from_this()));
+  }
+
+  void OnClose(beast::error_code ec) {
+    if (ec) {
+      return Fail(ec, "close");
+    }
+  }
+
+ private:
+  using InternalStream = std::conditional_t<ssl, beast::ssl_stream<beast::tcp_stream>, beast::tcp_stream>;
+  tcp::resolver resolver_;
+  websocket::stream<InternalStream> ws_;
+  beast::flat_buffer buffer_;
+  std::string host_;
+  std::vector<std::string> &received_messages_;
+  std::optional<Credentials> creds_{std::nullopt};
+};
 
 std::unique_ptr<mg::Client> GetBoltClient(const uint16_t bolt_port, const bool use_ssl) {
   auto client = mg::Client::Connect({.host = "127.0.0.1", .port = bolt_port, .use_ssl = use_ssl});
@@ -107,7 +241,7 @@ inline void AssertLogMessage(const std::string &log_message) {
   MG_ASSERT(json_message.at("event").is_string(), "Event is not a string!");
   MG_ASSERT(json_message.at("event").get<std::string>() == "log", "Event is not equal to `log`!");
   MG_ASSERT(json_message.at("level").is_string(), "Level is not a string!");
-  MG_ASSERT(std::ranges::count(kSupportedLevels, json_message.at("level")) > 0);
+  MG_ASSERT(std::ranges::count(kSupportedLevels, json_message.at("level")) == 1);
   MG_ASSERT(json_message.at("message").is_string(), "Message is not a string!");
 }
 
@@ -156,7 +290,7 @@ template <typename TWebsocketClient>
 void TestWebsocketWithoutBeingAuthorized(std::unique_ptr<mg::Client> &mg_client) {
   spdlog::info("Starting websocket connection with users but without being authenticated.");
   std::this_thread::sleep_for(std::chrono::seconds(1));
-  auto websocket_client = TWebsocketClient();
+  auto websocket_client = TWebsocketClient({"wrong", "credentials"});
   websocket_client.Connect("127.0.0.1", "7444");
 
   RunQueries(mg_client);
@@ -166,7 +300,7 @@ void TestWebsocketWithoutBeingAuthorized(std::unique_ptr<mg::Client> &mg_client)
   const auto received_messages = websocket_client.GetReceivedMessages();
   spdlog::info("Received {} messages.", received_messages.size());
 
-  MG_ASSERT(received_messages.size() < 2, "There is no more than one received message!");
+  MG_ASSERT(received_messages.size() == 1, "There must be only one message received!");
   if (!received_messages.empty()) {
     auto json_message = nlohmann::json::parse(received_messages[0]);
     AssertAuthMessage(json_message, false);
