+++ tests/e2e/websocket/workloads.yaml
@@ -1,5 +1,6 @@
 cert_file: &cert_file "$PROJECT_DIR/tests/e2e/websocket/memgraph-selfsigned.crt"
 key_file: &key_file "$PROJECT_DIR/tests/e2e/websocket/memgraph-selfsigned.key"
+bolt_port: &bolt_port "7687"
 template_cluster: &template_cluster
   cluster:
     websocket:
@@ -10,7 +11,8 @@ template_cluster_ssl: &template_cluster_ssl
     websocket:
       args:
         [
-          "--bolt-port=7686",
+          "--bolt-port",
+          *bolt_port,
           "--log-level=TRACE",
           "--bolt-cert-file",
           *cert_file,
@@ -24,9 +26,9 @@ template_cluster_ssl: &template_cluster_ssl
 workloads:
   - name: "Websocket"
     binary: "tests/e2e/websocket/memgraph__e2e__websocket"
-    args: ["--bolt-port", "7687"]
+    args: ["--bolt-port", *bolt_port]
     <<: *template_cluster
   - name: "Websocket SSL"
     binary: "tests/e2e/websocket/memgraph__e2e__websocket_ssl"
-    args: ["--bolt-port", "7686"]
+    args: ["--bolt-port", *bolt_port]
     <<: *template_cluster_ssl
